+++ packages/angular-workspace/spright-angular/chat/input/spright-chat-input.directive.ts
@@ -31,6 +31,22 @@ export class SprightChatInputDirective {
         this.renderer.setProperty(this.elementRef.nativeElement, 'sendButtonLabel', value);
     }
 
+    public get stopButtonLabel(): string | undefined {
+        return this.elementRef.nativeElement.stopButtonLabel;
+    }
+
+    @Input('stop-button-label') public set stopButtonLabel(value: string | undefined) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'stopButtonLabel', value);
+    }
+
+    public get processing(): boolean | undefined {
+        return this.elementRef.nativeElement.processing;
+    }
+
+    @Input('processing') public set processing(value: boolean | undefined) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'processing', value);
+    }
+
     public get maxLength(): number | undefined {
         return this.elementRef.nativeElement.maxLength;
     }