+++ packages/angular-workspace/spright-angular/chat/input/tests/spright-chat-input.directive.spec.ts
@@ -54,6 +54,16 @@ describe('Spright chat input', () => {
             expect(nativeElement.sendButtonLabel).toBeUndefined();
         });
 
+        it('has expected defaults for stopButtonLabel', () => {
+            expect(directive.stopButtonLabel).toBeUndefined();
+            expect(nativeElement.stopButtonLabel).toBeUndefined();
+        });
+
+        it('has expected defaults for processing', () => {
+            expect(directive.processing).toBeFalse();
+            expect(nativeElement.processing).toBeFalse();
+        });
+
         it('has expected defaults for value', () => {
             expect(directive.value).toEqual('');
             expect(nativeElement.value).toEqual('');
@@ -81,6 +91,8 @@ describe('Spright chat input', () => {
                 <spright-chat-input #chatInput
                     placeholder="Placeholder value"
                     send-button-label="Send button label value"
+                    stop-button-label="Stop button label value"
+                    processing="true"
                     value="Value value"
                     maxlength="10"
                     error-text="Error text value"
@@ -119,6 +131,16 @@ describe('Spright chat input', () => {
             expect(nativeElement.sendButtonLabel).toBe('Send button label value');
         });
 
+        it('will use template string values for stopButtonLabel', () => {
+            expect(directive.stopButtonLabel).toBe('Stop button label value');
+            expect(nativeElement.stopButtonLabel).toBe('Stop button label value');
+        });
+
+        it('will use template string values for processing', () => {
+            expect(directive.processing).toBeTrue();
+            expect(nativeElement.processing).toBeTrue();
+        });
+
         it('will use template string values for value', () => {
             expect(directive.value).toBe('Value value');
             expect(nativeElement.value).toBe('Value value');
@@ -146,6 +168,8 @@ describe('Spright chat input', () => {
                 <spright-chat-input #chatInput
                     [placeholder]="placeholder"
                     [sendButtonLabel]="sendButtonLabel"
+                    [stopButtonLabel]="stopButtonLabel"
+                    [processing]="processing"
                     [value]="value"
                     [maxlength]="maxLength"
                     [errorText]="errorText"
@@ -159,7 +183,9 @@ describe('Spright chat input', () => {
             @ViewChild('chatInput', { read: ElementRef }) public elementRef: ElementRef<ChatInput>;
             public placeholder = 'initial';
             public sendButtonLabel = 'initial';
+            public stopButtonLabel = 'initial';
             public value = 'initial';
+            public processing = false;
             public maxLength = 20;
             public errorText = 'initial';
             public errorVisible: BooleanValueOrAttribute = null;
@@ -201,6 +227,28 @@ describe('Spright chat input', () => {
 
             expect(directive.sendButtonLabel).toBe('updated sendButtonLabel value');
             expect(nativeElement.sendButtonLabel).toBe('updated sendButtonLabel value');
+        });
+
+        it('can be configured with property binding for stopButtonLabel', () => {
+            expect(directive.stopButtonLabel).toBe('initial');
+            expect(nativeElement.stopButtonLabel).toBe('initial');
+
+            fixture.componentInstance.stopButtonLabel = 'updated stopButtonLabel value';
+            fixture.detectChanges();
+
+            expect(directive.stopButtonLabel).toBe('updated stopButtonLabel value');
+            expect(nativeElement.stopButtonLabel).toBe('updated stopButtonLabel value');
+        });
+
+        it('can be configured with property binding for processing', () => {
+            expect(directive.processing).toBeFalse();
+            expect(nativeElement.processing).toBeFalse();
+
+            fixture.componentInstance.processing = true;
+            fixture.detectChanges();
+
+            expect(directive.processing).toBeTrue();
+            expect(nativeElement.processing).toBeTrue();
         });
 
         it('can be configured with property binding for value', () => {
@@ -254,6 +302,8 @@ describe('Spright chat input', () => {
                 <spright-chat-input #chatInput
                     [attr.placeholder]="placeholder"
                     [attr.send-button-label]="sendButtonLabel"
+                    [attr.stop-button-label]="stopButtonLabel"
+                    [attr.processing]="processing"
                     [attr.value]="value"
                     [attr.maxlength]="maxLength"
                     [attr.error-text]="errorText"
@@ -267,6 +317,8 @@ describe('Spright chat input', () => {
             @ViewChild('chatInput', { read: ElementRef }) public elementRef: ElementRef<ChatInput>;
             public placeholder = 'initial';
             public sendButtonLabel = 'initial';
+            public stopButtonLabel = 'initial';
+            public processing = false;
             public value = 'initial';
             public maxLength = 20;
             public errorText = 'initial';
@@ -308,6 +360,28 @@ describe('Spright chat input', () => {
 
             expect(directive.sendButtonLabel).toBe('updated sendButtonLabel value');
             expect(nativeElement.sendButtonLabel).toBe('updated sendButtonLabel value');
+        });
+
+        it('can be configured with attribute binding for stopButtonLabel', () => {
+            expect(directive.stopButtonLabel).toBe('initial');
+            expect(nativeElement.stopButtonLabel).toBe('initial');
+
+            fixture.componentInstance.stopButtonLabel = 'updated stopButtonLabel value';
+            fixture.detectChanges();
+
+            expect(directive.stopButtonLabel).toBe('updated stopButtonLabel value');
+            expect(nativeElement.stopButtonLabel).toBe('updated stopButtonLabel value');
+        });
+
+        it('can be configured with attribute binding for processing', () => {
+            expect(directive.processing).toBe(false);
+            expect(nativeElement.processing).toBe(false);
+
+            fixture.componentInstance.processing = true;
+            fixture.detectChanges();
+
+            expect(directive.processing).toBe(true);
+            expect(nativeElement.processing).toBe(true);
         });
 
         it('can be configured with attribute binding for value', () => {