+++ packages/spright-components/src/chat/input/testing/chat-input.pageobject.ts
@@ -1,5 +1,7 @@
 import { keyEnter } from '@ni/fast-web-utilities';
 import { Button } from '@ni/nimble-components/dist/esm/button';
+import { iconPaperPlaneTag } from '@ni/nimble-components/dist/esm/icons/paper-plane';
+import { iconStopSquareTag } from '@ni/nimble-components/dist/esm/icons/stop-square';
 import {
     processUpdates,
     waitForUpdatesAsync
@@ -14,8 +16,12 @@ import type { ChatInput } from '..';
 export class ChatInputPageObject {
     public constructor(protected readonly element: ChatInput) {}
 
-    public isSendButtonEnabled(): boolean {
-        return !this.getSendButton().disabled;
+    public isButtonEnabled(): boolean {
+        return !this.getActionButton().disabled;
+    }
+
+    public isProcessing(): boolean {
+        return this.element.processing;
     }
 
     public isTextAreaFocused(): boolean {
@@ -23,19 +29,35 @@ export class ChatInputPageObject {
     }
 
     public clickSendButton(): void {
-        this.getSendButton().click();
+        if (!this.element.processing) {
+            this.getActionButton().click();
+        }
     }
 
-    public getSendButtonTitle(): string {
-        return this.getSendButton().title;
+    public clickStopButton(): void {
+        if (this.element.processing) {
+            this.getActionButton().click();
+        }
     }
 
-    public getSendButtonTextContent(): string {
-        return this.getSendButton().textContent?.trim() ?? '';
+    public getButtonTitle(): string {
+        return this.getActionButton().title;
     }
 
-    public getSendButtonTabIndex(): string | null {
-        return this.getSendButton().getAttribute('tabindex');
+    public buttonHasSendIcon(): boolean {
+        return this.getActionButton().querySelector(iconPaperPlaneTag) !== null;
+    }
+
+    public buttonHasStopIcon(): boolean {
+        return this.getActionButton().querySelector(iconStopSquareTag) !== null;
+    }
+
+    public getButtonTextContent(): string {
+        return this.getActionButton().textContent?.trim() ?? '';
+    }
+
+    public getButtonTabIndex(): string | null {
+        return this.getActionButton().getAttribute('tabindex');
     }
 
     public textAreaHasFocus(): boolean {
@@ -81,9 +103,9 @@ export class ChatInputPageObject {
         await this.sendEnterKeyEvents(true);
     }
 
-    private getSendButton(): Button {
-        const sendButton = this.element.shadowRoot!.querySelector<Button>('.send-button')!;
-        return sendButton;
+    private getActionButton(): Button {
+        const actionButton = this.element.shadowRoot!.querySelector<Button>('.action-button')!;
+        return actionButton;
     }
 
     private async sendEnterKeyEvents(shiftKey: boolean): Promise<void> {